+++ sapao-na-cozinha/package.json
@@ -14,6 +14,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-router-dom": "^7.3.0",
+    "react-scroll": "^1.9.3",
     "tailwind-scrollbar": "^4.0.1",
     "tailwindcss": "^4.0.11"
   },
