+++ sapao-na-cozinha/src/components/pages/main-page/Footer.jsx
@@ -1,6 +1,6 @@
 export default function Footer() {
     return (
-      <footer className="bg-[var(--color-hunter-green)] text-[#F4F1E1] flex flex-col gap-2 items-center py-4">
+      <footer className="bg-[var(--color-dark-green)] text-[#F4F1E1] flex flex-col gap-2 items-center py-4">
         <p className="text-lg font-bold flex items-center">Feito com  <span className="material-symbols-outlined ml-[5px] mr-[5px]"> 
             favorite </span>  para quem ama cozinhar!</p>
         <p className="text-sm mt-2">© 2025 Sapão na Cozinha. Todos os direitos reservados.</p>
