+++ sapao-na-cozinha/src/components/pages/main-page/MainPage.jsx
@@ -1,17 +1,27 @@
+import { Element } from 'react-scroll';
+
 import PaginaInicial from './PaginaInicial';
 import PaginaCadastrar from './Cadastrar/PaginaCadastrar';
 import Header from './Header';
 import Footer from './Footer';
-import Cards from './Cards.jsx';
+import Cards from './Receitas/Cards';
 
 export default function MainPage() {
     return (
         <div>
             <Header />
-            <PaginaInicial />
-            <Cards />
-            <PaginaCadastrar/>
-            <Footer />
-        </div>
-    )
+            
+            <Element name='inicio' className='min-h-screen'>
+                <PaginaInicial />
+            </Element>
+            <Element name='cadastrar' className='min-h-screen'>
+                <PaginaCadastrar />
+            </Element>
+            <Element name='receitas' className='min-h-screen'>
+                <Cards />
+            </Element>
+
+            <Footer/>
+      </div>
+    );
 }
